requirements.txt
# Core dbt packages
dbt-core==1.5.0
dbt-postgres==1.5.0

# Database drivers
psycopg2-binary==2.9.6

# Airflow providers
apache-airflow-providers-amazon==8.7.0
apache-airflow-providers-postgres==5.7.0

# Required dependencies
python-dateutil==2.8.2
pandas==1.5.3  # Often needed for data transformations